requirements.txt
fastapi==0.78.0
python-telegram=0.15.0
PyYAML==6.0
uvicorn==0.17.6